Initial testing setup for the script
