+++ Assets/Scripts/NetworkTurnOn.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 54a7cb9b428a6334bb269e24234a084f
+timeCreated: 1453286703
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
