+++ Assets/Scripts/PlayerID.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0a604a5c856d5984ebe68fbb3ed1b7d3
+timeCreated: 1453286703
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
