+++ Assets/Scripts/RTTDebug.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6ed1b6d7e4a51054796158ffe68066d0
+timeCreated: 1453287890
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
