--- Assets/Scripts/WBFirstPersonController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 0ab17209d6bac0b4bb57938c8ad58bd4
-timeCreated: 1453240546
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
