--- Assets/Standard Assets/Characters/FirstPersonCharacter/Prefabs/FPSController.prefab.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 5e9e851c0e142814dac026a256ba2ac0
-NativeFormatImporter:
-  userData: 
-  assetBundleName:
