+++ Build_Data/output_log.txt
@@ -1,0 +1,78 @@
+Initialize engine version: 5.3.1f1 (cc9cbbcc37b4)
+GfxDevice: creating device client; threaded=1
+Direct3D:
+    Version:  Direct3D 11.0 [level 11.0]
+    Renderer: AMD Radeon HD 5700 Series (ID=0x68b8)
+    Vendor:   ATI
+    VRAM:     1010 MB
+Begin MonoManager ReloadAssembly
+Platform assembly: G:\UnityProjects\UnetCharacterController\Build_Data\Managed\UnityEngine.dll (this message is harmless)
+Loading G:\UnityProjects\UnetCharacterController\Build_Data\Managed\UnityEngine.dll into Unity Child Domain
+Platform assembly: G:\UnityProjects\UnetCharacterController\Build_Data\Managed\Assembly-CSharp-firstpass.dll (this message is harmless)
+Loading G:\UnityProjects\UnetCharacterController\Build_Data\Managed\Assembly-CSharp-firstpass.dll into Unity Child Domain
+Platform assembly: G:\UnityProjects\UnetCharacterController\Build_Data\Managed\Assembly-CSharp.dll (this message is harmless)
+Loading G:\UnityProjects\UnetCharacterController\Build_Data\Managed\Assembly-CSharp.dll into Unity Child Domain
+Platform assembly: G:\UnityProjects\UnetCharacterController\Build_Data\Managed\UnityEngine.UI.dll (this message is harmless)
+Loading G:\UnityProjects\UnetCharacterController\Build_Data\Managed\UnityEngine.UI.dll into Unity Child Domain
+Platform assembly: G:\UnityProjects\UnetCharacterController\Build_Data\Managed\UnityEngine.Networking.dll (this message is harmless)
+Loading G:\UnityProjects\UnetCharacterController\Build_Data\Managed\UnityEngine.Networking.dll into Unity Child Domain
+- Completed reload, in  0.050 seconds
+Platform assembly: G:\UnityProjects\UnetCharacterController\Build_Data\Managed\System.Core.dll (this message is harmless)
+Platform assembly: G:\UnityProjects\UnetCharacterController\Build_Data\Managed\System.dll (this message is harmless)
+desktop: 1920x1080 60Hz; virtual: 3200x1080 at -1280,0
+<RI> Initializing input.
+<RI> Input initialized.
+<RI> Initialized touch support.
+UnloadTime: 0.735682 ms
+Setting up 1 worker threads for Enlighten.
+  Thread -> id: 49ec -> priority: 1 
+network simulation allowed only for editor or developer build or in debug mode only
+ 
+(Filename:  Line: 511)
+
+network simulation allowed only for editor or developer build or in debug mode only
+ 
+(Filename:  Line: 511)
+
+Socket: bind failed, error: Only one usage of each socket address (protocol/network address/port) is normally permitted.
+Cannot open socket on ip {*} and port {7777}; check please your network, most probably port has been already occupied
+ 
+(Filename:  Line: 488)
+
+StartServer listen failed.
+ 
+(Filename: C:/buildslave/unity/build/artifacts/generated/common/runtime/UnityEngineDebugBindings.gen.cpp Line: 37)
+
+Socket: bind failed, error: Only one usage of each socket address (protocol/network address/port) is normally permitted.
+Cannot open socket on ip {*} and port {7777}; check please your network, most probably port has been already occupied
+ 
+(Filename:  Line: 488)
+
+StartServer listen failed.
+ 
+(Filename: C:/buildslave/unity/build/artifacts/generated/common/runtime/UnityEngineDebugBindings.gen.cpp Line: 37)
+
+Did not find target for RPC message for 2
+ 
+(Filename: C:/buildslave/unity/build/artifacts/generated/common/runtime/UnityEngineDebugBindings.gen.cpp Line: 37)
+
+Did not find target for RPC message for 6
+ 
+(Filename: C:/buildslave/unity/build/artifacts/generated/common/runtime/UnityEngineDebugBindings.gen.cpp Line: 37)
+
+Did not find target for RPC message for 7
+ 
+(Filename: C:/buildslave/unity/build/artifacts/generated/common/runtime/UnityEngineDebugBindings.gen.cpp Line: 37)
+
+Did not find target for RPC message for 8
+ 
+(Filename: C:/buildslave/unity/build/artifacts/generated/common/runtime/UnityEngineDebugBindings.gen.cpp Line: 37)
+
+Did not find target for RPC message for 9
+ 
+(Filename: C:/buildslave/unity/build/artifacts/generated/common/runtime/UnityEngineDebugBindings.gen.cpp Line: 37)
+
+Did not find target for RPC message for 10
+ 
+(Filename: C:/buildslave/unity/build/artifacts/generated/common/runtime/UnityEngineDebugBindings.gen.cpp Line: 37)
+
